+++ front/src/hooks/auth/useSignInAnonymously.ts
@@ -1,33 +1,16 @@
 import { useRouter } from 'next/router'
 import { getAuth, onAuthStateChanged, signInAnonymously } from 'firebase/auth'
-import {
-  getFirestore,
-  collection,
-  doc,
-  getDoc,
-  setDoc
-} from 'firebase/firestore'
 import { useSetRecoilState } from 'recoil'
 import { useBoolean } from '@chakra-ui/react'
-import { CurrentUser } from '@/types/db'
 import { currentUserState } from '@/recoil/atoms'
+import { useCreateUser } from '@/hooks/firestore'
 
 export const useSignInAnonymously = () => {
   const [disabled, setDisabled] = useBoolean()
   const router = useRouter()
   const setCurrentUser = useSetRecoilState(currentUserState)
 
-  const createUserIfNotFound = async (user: CurrentUser) => {
-    const db = getFirestore()
-    const usersCollection = collection(db, 'users')
-    const userRef = doc(usersCollection, user.uid)
-    const document = await getDoc(userRef)
-    if (document.exists()) return
-    await setDoc(userRef, {
-      uid: user.uid,
-      isAnonymous: user.isAnonymous
-    })
-  }
+  const createUser = useCreateUser
 
   const signInAnonymouslyHandler = async () => {
     setDisabled.on()
@@ -41,11 +24,12 @@ export const useSignInAnonymously = () => {
             uid: currentUser.uid,
             isAnonymous: currentUser.isAnonymous
           })
-          createUserIfNotFound(currentUser)
+          createUser(currentUser)
         } else {
           setCurrentUser(null)
         }
       })
+
       router.push('/user')
     } catch (err) {
       console.error(err)
