+++ front/src/hooks/firestore/user/useCreateUser.ts
@@ -1,0 +1,22 @@
+import { User } from 'firebase/auth'
+import {
+  getDoc,
+  doc,
+  collection,
+  getFirestore,
+  setDoc
+} from 'firebase/firestore'
+
+export const useCreateUser = async (user: User) => {
+  const db = getFirestore()
+  const usersCollection = collection(db, 'users')
+  const userRef = doc(usersCollection, user.uid)
+  const document = await getDoc(userRef)
+
+  if (document.exists()) return
+
+  await setDoc(userRef, {
+    uid: user.uid,
+    isAnonymous: user.isAnonymous
+  })
+}
